archive/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "The backend for the LLMs for Science project."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3",
    "uv",
    "Flask",
    "psycopg[binary]",
    "uuid7",
    "psycopg_pool",
    "blinker",
    "click",
    "colorama",
    "itsdangerous",
    "Jinja2",
    "MarkupSafe",
    "python-dotenv",
    "Werkzeug",
    "requests>=2.32.3",
    "faker>=36.1.1",
    "gunicorn>=21.2.0",
    "transformers>=4.49.0",
    "pymupdf>=1.25.3",
    "ollama>=0.4.7",
    "flask-cors>=5.0.0",
    "arxiv>=2.1.3",
    "texsoup>=0.3.1",
    "pandoc>=2.4",
    "openai>=1.69.0",
    "pdfreader>=0.1.15",
    "instructor>=1.7.7",
    "unstructured[pdf]>=0.17.2",
    "pdfminer.six>=20221105",
    "pi-heif>=0.13.0",
    "unstructured-inference>=0.7.24",
    "pdf2image>=1.17.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
    packages = ["./"]

[dependency-groups]
dev = [
    "faker>=36.1.1",
    "pytest>=8.3.4",
    "ruff>=0.9.6",
    "ipython>=8.32.0",
    "ipykernel>=6.29.5",
]


[tool.pytest.ini_options]
minversion = "8.3.4"
addopts = "-ra -v"
testpaths = [
    "tests",
]
filterwarnings = [
    "ignore:builtin type SwigPyPacked has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type SwigPyObject has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type swigvarlink has no __module__ attribute:DeprecationWarning",
    "ignore:read_text is deprecated:DeprecationWarning",
    "ignore:open_text is deprecated:DeprecationWarning",
    "ignore:min_items is deprecated:DeprecationWarning"
]
